Cleanup and changes to help command, and '/awf world'
Stopped sending the 'awf-tip' to the console when the help command is run.

Switched the italicized 'world' to '<world>' for added clarity and for better looks in the console.

Added myself as a maintainer in the info command an in the main class.

Changed the plugin website and added myself to the authors list in plugin.yml

Added the command '/awf world'. Given it is an added feature I should increment the version to 2.4, however, since I will be replacing this at a later date I am not going to do that.

Added crappy translations for '/awf world'
I am at least moderately familiar with Spanish so hopefully those translations are better than the ones done by Google Translate.