1.17 support
I apologize for this taking so long to update. I've been busy with life things and I didn't have the time to update MCShared.

Sidpatchy/MCShared is no longer using NMS because I was dumb, however, me being dumb lead to a good thing so I'll call it intentional.

Sequence of events:
1. MCShared was using NMS to achieve hoverable chat messages
2. It broke because of Spigot changes to 1.17
3. I fixed it by copying Pyves' changes from AdvancedAchievements (didn't realize I fixed it)
4. My build settings were messed up on IntelliJ and I was still building with the old version of MCShared. (didn't realize my build settings were messed up)
5. I rewrote the entire "FancyMessageSender" class to use Bungee's chat API instead of NMS.
6. I figured out that 4 was an issue
7. I fixed my build settings

See: #24